Split the sql generating functions into a separate function.
